Added completion messages, fixed up copy-root-to-root.sh, and more
Added completion messages such as "Successfully installed the programs! ", expanded on the copy-root-to-root.sh functionality to pull files from all-distros and ubuntu-only (going to add arch-only in the future) and not having a copy of every file in root/etc/skel, and added a script to fix the Steam runtime on Radeon 6520G and possibly more GPUs.